==> 1. final_79_mod.R <==

# Set working directory
# setwd()

data_79 <- read.table('./data/NLSY79/final_79/final_79.dat', sep=' ')
names(data_79) <- c('R0000100',
                     'R0000300',
                     'R0000500',
                     'R0000700',
                     'R0017300',
                     'R0046400',
                     'R0155400',
                     'R0155500',
                     'R0156000',
                     'R0156100',
                     'R0169100',
                     'R0173600',
                     'R0214700',
                     'R0214800',
                     'R0214900',
                     'R0214901',
                     'R0216101',
                     'R0217900',
                     'R0218001',
                     'R0229200',
                     'R0263400',
                     'R0312300',
                     'R0312600',
                     'R0312710',
                     'R0313000',
                     'R0328000',
                     'R0405201',
                     'R0406010',
                     'R0406300',
                     'R0406301',
                     'R0407601',
                     'R0417400',
                     'R0446400',
                     'R0482600',
                     'R0482910',
                     'R0483200',
                     'R0483500',
                     'R0498500',
                     'R0614601',
                     'R0618410',
                     'R0618800',
                     'R0618801',
                     'R0647101',
                     'R0664500',
                     'R0702100',
                     'R0782100',
                     'R0782101',
                     'R0782400',
                     'R0782401',
                     'R0784300',
                     'R0784301',
                     'R0784600',
                     'R0784601',
                     'R0798600',
                     'R0828000',
                     'R0896701',
                     'R0898500',
                     'R0898501',
                     'R0898600',
                     'R0898838',
                     'R0905900',
                     'R0945000',
                     'R0945300',
                     'R1024000',
                     'R1024001',
                     'R1024300',
                     'R1024301',
                     'R1026200',
                     'R1026201',
                     'R1026500',
                     'R1026501',
                     'R1144401',
                     'R1144500',
                     'R1144700',
                     'R1144701',
                     'R1146830',
                     'R1205800',
                     'R1255400',
                     'R1255700',
                     'R1410700',
                     'R1410701',
                     'R1411000',
                     'R1411001',
                     'R1412900',
                     'R1412901',
                     'R1413200',
                     'R1413201',
                     'R1519601',
                     'R1519700',
                     'R1519900',
                     'R1519901',
                     'R1522037',
                     'R1605100',
                     'R1650200',
                     'R1650500',
                     'R1778500',
                     'R1778501',
                     'R1778800',
                     'R1778801',
                     'R1780700',
                     'R1780701',
                     'R1781000',
                     'R1781001',
                     'R1890201',
                     'R1890400',
                     'R1890600',
                     'R1890601',
                     'R1892737',
                     'R1905600',
                     'R1922800',
                     'R1923100',
                     'R2141600',
                     'R2141601',
                     'R2141900',
                     'R2141901',
                     'R2143800',
                     'R2143801',
                     'R2144100',
                     'R2144101',
                     'R2257301',
                     'R2257500',
                     'R2257700',
                     'R2257701',
                     'R2259837',
                     'R2306500',
                     'R2317600',
                     'R2317900',
                     'R2350300',
                     'R2350301',
                     'R2350600',
                     'R2350601',
                     'R2352500',
                     'R2352501',
                     'R2352800',
                     'R2352801',
                     'R2444501',
                     'R2444700',
                     'R2445100',
                     'R2445101',
                     'R2448037',
                     'R2509000',
                     'R2525400',
                     'R2525700',
                     'R2722500',
                     'R2722501',
                     'R2722800',
                     'R2722801',
                     'R2724700',
                     'R2724701',
                     'R2725000',
                     'R2725001',
                     'R2870001',
                     'R2870200',
                     'R2870600',
                     'R2870700',
                     'R2877600',
                     'R2908100',
                     'R2924400',
                     'R2924700',
                     'R2971400',
                     'R2971401',
                     'R2971700',
                     'R2971701',
                     'R2973600',
                     'R2973601',
                     'R2973900',
                     'R2973901',
                     'R3073801',
                     'R3074000',
                     'R3074300',
                     'R3074400',
                     'R3076842',
                     'R3110200',
                     'R3127100',
                     'R3127400',
                     'R3279400',
                     'R3279401',
                     'R3279700',
                     'R3279701',
                     'R3281600',
                     'R3281601',
                     'R3281900',
                     'R3281901',
                     'R3292201',
                     'R3292501',
                     'R3400201',
                     'R3400700',
                     'R3401000',
                     'R3401100',
                     'R3407700',
                     'R3510200',
                     'R3522800',
                     'R3523100',
                     'R3559000',
                     'R3559001',
                     'R3559300',
                     'R3559301',
                     'R3561200',
                     'R3561201',
                     'R3561500',
                     'R3561501',
                     'R3571801',
                     'R3572101',
                     'R3655801',
                     'R3656100',
                     'R3656400',
                     'R3656500',
                     'R3659047',
                     'R3710200',
                     'R3727800',
                     'R3728100',
                     'R3897100',
                     'R3897101',
                     'R3897400',
                     'R3897401',
                     'R3899300',
                     'R3899301',
                     'R3899600',
                     'R3899601',
                     'R3909901',
                     'R3910201',
                     'R4006301',
                     'R4006600',
                     'R4006900',
                     'R4007000',
                     'R4009447',
                     'R4137900',
                     'R4182100',
                     'R4193700',
                     'R4295100',
                     'R4295101',
                     'R4295400',
                     'R4295500',
                     'R4295501',
                     'R4314400',
                     'R4314401',
                     'R4314900',
                     'R4314901',
                     'R4390801',
                     'R4391301',
                     'R4417401',
                     'R4417700',
                     'R4418000',
                     'R4418100',
                     'R4444700',
                     'R4526500',
                     'R4587901',
                     'R4982800',
                     'R4982801',
                     'R4983100',
                     'R4983200',
                     'R4983201',
                     'R4996000',
                     'R4996001',
                     'R4996600',
                     'R4996601',
                     'R5080401',
                     'R5080700',
                     'R5081000',
                     'R5081100',
                     'R5087500',
                     'R5165701',
                     'R5166000',
                     'R5166300',
                     'R5166400',
                     'R5172800',
                     'R5221800',
                     'R5270600',
                     'R5270900',
                     'R5626200',
                     'R5626201',
                     'R5626500',
                     'R5626600',
                     'R5626601',
                     'R5650800',
                     'R5650801',
                     'R5651400',
                     'R5651401',
                     'R5821800',
                     'R6364600',
                     'R6364601',
                     'R6364900',
                     'R6365000',
                     'R6365001',
                     'R6374900',
                     'R6374901',
                     'R6375300',
                     'R6375301',
                     'R6466301',
                     'R6473700',
                     'R6478700',
                     'R6478900',
                     'R6479000',
                     'R6486400',
                     'R6540400',
                     'R6592900',
                     'R6839600',
                     'R6909700',
                     'R6909701',
                     'R6910000',
                     'R6911100',
                     'R6911101',
                     'R6917800',
                     'R6917801',
                     'R6919200',
                     'R6919201',
                     'R7006201',
                     'R7006500',
                     'R7103600',
                     'R7209600',
                     'R7548600',
                     'R7607800',
                     'R7608200',
                     'R7608300',
                     'R7608400',
                     'R7608500',
                     'R7608800',
                     'R7609000',
                     'R7609300',
                     'R7609400',
                     'R7609500',
                     'R7609600',
                     'R7609700',
                     'R7617300',
                     'R7617600',
                     'R7617700',
                     'R7617800',
                     'R7617900',
                     'R7618000',
                     'R7618500',
                     'R7618800',
                     'R7618900',
                     'R7619000',
                     'R7619100',
                     'R7619200',
                     'R7703401',
                     'R7703700',
                     'R7729800',
                     'R7729900',
                     'R7730100',
                     'R7730200',
                     'R7810500',
                     'R7898000',
                     'R8255400',
                     'R8316300',
                     'R8316400',
                     'R8316500',
                     'R8316700',
                     'R8316800',
                     'R8317100',
                     'R8318200',
                     'R8318300',
                     'R8318400',
                     'R8318500',
                     'R8318600',
                     'R8318700',
                     'R8325800',
                     'R8326000',
                     'R8326200',
                     'R8326300',
                     'R8326400',
                     'R8326500',
                     'R8328000',
                     'R8328100',
                     'R8328200',
                     'R8328300',
                     'R8328400',
                     'R8328500',
                     'R8495800',
                     'R8496100',
                     'T0014400',
                     'T0138400',
                     'T0857100',
                     'T0912400',
                     'T0912600',
                     'T0912800',
                     'T0913000',
                     'T0913100',
                     'T0913400',
                     'T0913900',
                     'T0914100',
                     'T0914300',
                     'T0914400',
                     'T0914500',
                     'T0914600',
                     'T0920800',
                     'T0920900',
                     'T0921100',
                     'T0921200',
                     'T0921300',
                     'T0921400',
                     'T0922200',
                     'T0922300',
                     'T0922400',
                     'T0922500',
                     'T0922600',
                     'T0922700',
                     'T0987400',
                     'T0987800',
                     'T0988100',
                     'T0988200',
                     'T1214300',
                     'T1215600',
                     'T1298000',
                     'T2015100',
                     'T2076700',
                     'T2076900',
                     'T2077100',
                     'T2077300',
                     'T2077400',
                     'T2077700',
                     'T2078800',
                     'T2079100',
                     'T2079200',
                     'T2079300',
                     'T2085500',
                     'T2085700',
                     'T2085900',
                     'T2086000',
                     'T2086100',
                     'T2086200',
                     'T2087700',
                     'T2087800',
                     'T2088000',
                     'T2088100',
                     'T2088200',
                     'T2088300',
                     'T2209700',
                     'T2210000',
                     'T2272800',
                     'T2326500',
                     'T2998600',
                     'T3045300',
                     'T3045500',
                     'T3045700',
                     'T3046000',
                     'T3046100',
                     'T3046400',
                     'T3047500',
                     'T3047600',
                     'T3047700',
                     'T3047800',
                     'T3047900',
                     'T3048000',
                     'T3056000',
                     'T3056200',
                     'T3056400',
                     'T3056600',
                     'T3056700',
                     'T3056800',
                     'T3058300',
                     'T3058400',
                     'T3058500',
                     'T3058600',
                     'T3058700',
                     'T3058800',
                     'T3107500',
                     'T3107800',
                     'T3212900',
                     'T3308700',
                     'T3942500',
                     'T3977400',
                     'T3977500',
                     'T3977600',
                     'T3977900',
                     'T3978000',
                     'T3978300',
                     'T3979400',
                     'T3979500',
                     'T3979600',
                     'T3979700',
                     'T3979800',
                     'T3979900',
                     'T3987600',
                     'T3987800',
                     'T3988000',
                     'T3988200',
                     'T3988300',
                     'T3988400',
                     'T3989900',
                     'T3990000',
                     'T3990100',
                     'T3990200',
                     'T3990300',
                     'T3990400',
                     'T4112000',
                     'T4112300',
                     'T4201100',
                     'T4282800',
                     'T4876800',
                     'T4915800',
                     'T4915900',
                     'T4916000',
                     'T4916300',
                     'T4916400',
                     'T4916700',
                     'T4917800',
                     'T4917900',
                     'T4918000',
                     'T4918100',
                     'T4918200',
                     'T4918300',
                     'T4924900',
                     'T4925100',
                     'T4925300',
                     'T4925500',
                     'T4925600',
                     'T4925700',
                     'T4927200',
                     'T4927300',
                     'T4927400',
                     'T4927500',
                     'T4927600',
                     'T4927700',
                     'T5022200',
                     'T5022600',
                     'T5176100',
                     'T5256900',
                     'T5593000',
                     'T5619500',
                     'T5619700',
                     'T5619900',
                     'T5620200',
                     'T5620300',
                     'T5620600',
                     'T5621700',
                     'T5621800',
                     'T5622000',
                     'T5622100',
                     'T5622200',
                     'T5622300',
                     'T5630100',
                     'T5630300',
                     'T5630500',
                     'T5630700',
                     'T5630800',
                     'T5630900',
                     'T5632400',
                     'T5632800',
                     'T5632900',
                     'T5633000',
                     'T5633100',
                     'T5770500',
                     'T5770800')


# Handle missing values

data_79[data_79 == -1] = NA  # Refused 
data_79[data_79 == -2] = NA  # Dont know 
data_79[data_79 == -3] = NA  # Invalid missing 
data_79[data_79 == -4] = NA  # Valid missing 
data_79[data_79 == -5] = NA  # Non-interview
