==> 2. final_97_mod.R <==

# Set working directory
# setwd()


data_97 <- read.table('./data/NLSY97/final_97/final_97.dat', sep=' ')
names(data_97) <- c('E0021901',
                     'E0021902',
                     'E0021903',
                     'E0021904',
                     'E0021905',
                     'E0021906',
                     'E0021907',
                     'E0021908',
                     'E0021909',
                     'E0021910',
                     'E0021911',
                     'E0021912',
                     'E0021913',
                     'E0021914',
                     'E0021915',
                     'E0021916',
                     'E0021917',
                     'E0021918',
                     'E0021919',
                     'E0021920',
                     'E0021921',
                     'E0021922',
                     'E0021923',
                     'E0021924',
                     'E0021925',
                     'E0021926',
                     'E0021927',
                     'E0021928',
                     'E0021929',
                     'E0021930',
                     'E0021931',
                     'E0021932',
                     'E0021933',
                     'E0021934',
                     'E0021935',
                     'E0021936',
                     'E0021937',
                     'E0021938',
                     'E0021939',
                     'E0021940',
                     'E0021941',
                     'E0021942',
                     'E0021943',
                     'E0021944',
                     'E0021945',
                     'E0021946',
                     'E0021947',
                     'E0021948',
                     'E0021949',
                     'E0021950',
                     'E0021951',
                     'E0021952',
                     'E0022001',
                     'E0022002',
                     'E0022003',
                     'E0022004',
                     'E0022005',
                     'E0022006',
                     'E0022007',
                     'E0022008',
                     'E0022009',
                     'E0022010',
                     'E0022011',
                     'E0022012',
                     'E0022013',
                     'E0022014',
                     'E0022015',
                     'E0022016',
                     'E0022017',
                     'E0022018',
                     'E0022019',
                     'E0022020',
                     'E0022021',
                     'E0022022',
                     'E0022023',
                     'E0022024',
                     'E0022025',
                     'E0022026',
                     'E0022027',
                     'E0022028',
                     'E0022029',
                     'E0022030',
                     'E0022031',
                     'E0022032',
                     'E0022033',
                     'E0022034',
                     'E0022035',
                     'E0022036',
                     'E0022037',
                     'E0022038',
                     'E0022039',
                     'E0022040',
                     'E0022041',
                     'E0022042',
                     'E0022043',
                     'E0022044',
                     'E0022045',
                     'E0022046',
                     'E0022047',
                     'E0022048',
                     'E0022049',
                     'E0022050',
                     'E0022051',
                     'E0022052',
                     'E0022053',
                     'E0022101',
                     'E0022102',
                     'E0022103',
                     'E0022104',
                     'E0022105',
                     'E0022106',
                     'E0022107',
                     'E0022108',
                     'E0022109',
                     'E0022110',
                     'E0022111',
                     'E0022112',
                     'E0022113',
                     'E0022114',
                     'E0022115',
                     'E0022116',
                     'E0022117',
                     'E0022118',
                     'E0022119',
                     'E0022120',
                     'E0022121',
                     'E0022122',
                     'E0022123',
                     'E0022124',
                     'E0022125',
                     'E0022126',
                     'E0022127',
                     'E0022128',
                     'E0022129',
                     'E0022130',
                     'E0022131',
                     'E0022132',
                     'E0022133',
                     'E0022134',
                     'E0022135',
                     'E0022136',
                     'E0022137',
                     'E0022138',
                     'E0022139',
                     'E0022140',
                     'E0022141',
                     'E0022142',
                     'E0022143',
                     'E0022144',
                     'E0022145',
                     'E0022146',
                     'E0022147',
                     'E0022148',
                     'E0022149',
                     'E0022150',
                     'E0022151',
                     'E0022152',
                     'E0022201',
                     'E0022202',
                     'E0022203',
                     'E0022204',
                     'E0022205',
                     'E0022206',
                     'E0022207',
                     'E0022208',
                     'E0022209',
                     'E0022210',
                     'E0022211',
                     'E0022212',
                     'E0022213',
                     'E0022214',
                     'E0022215',
                     'E0022216',
                     'E0022217',
                     'E0022218',
                     'E0022219',
                     'E0022220',
                     'E0022221',
                     'E0022222',
                     'E0022223',
                     'E0022224',
                     'E0022225',
                     'E0022226',
                     'E0022227',
                     'E0022228',
                     'E0022229',
                     'E0022230',
                     'E0022231',
                     'E0022232',
                     'E0022233',
                     'E0022234',
                     'E0022235',
                     'E0022236',
                     'E0022237',
                     'E0022238',
                     'E0022239',
                     'E0022240',
                     'E0022241',
                     'E0022242',
                     'E0022243',
                     'E0022244',
                     'E0022245',
                     'E0022246',
                     'E0022247',
                     'E0022248',
                     'E0022249',
                     'E0022250',
                     'E0022251',
                     'E0022252',
                     'E0022301',
                     'E0022302',
                     'E0022303',
                     'E0022304',
                     'E0022305',
                     'E0022306',
                     'E0022307',
                     'E0022308',
                     'E0022309',
                     'E0022310',
                     'E0022311',
                     'E0022312',
                     'E0022313',
                     'E0022314',
                     'E0022315',
                     'E0022316',
                     'E0022317',
                     'E0022318',
                     'E0022319',
                     'E0022320',
                     'E0022321',
                     'E0022322',
                     'E0022323',
                     'E0022324',
                     'E0022325',
                     'E0022326',
                     'E0022327',
                     'E0022328',
                     'E0022329',
                     'E0022330',
                     'E0022331',
                     'E0022332',
                     'E0022333',
                     'E0022334',
                     'E0022335',
                     'E0022336',
                     'E0022337',
                     'E0022338',
                     'E0022339',
                     'E0022340',
                     'E0022341',
                     'E0022342',
                     'E0022343',
                     'E0022344',
                     'E0022345',
                     'E0022346',
                     'E0022347',
                     'E0022348',
                     'E0022349',
                     'E0022350',
                     'E0022351',
                     'E0022352',
                     'E0022401',
                     'E0022402',
                     'E0022403',
                     'E0022404',
                     'E0022405',
                     'E0022406',
                     'E0022407',
                     'E0022408',
                     'E0022409',
                     'E0022410',
                     'E0022411',
                     'E0022412',
                     'E0022413',
                     'E0022414',
                     'E0022415',
                     'E0022416',
                     'E0022417',
                     'E0022418',
                     'E0022419',
                     'E0022420',
                     'E0022421',
                     'E0022422',
                     'E0022423',
                     'E0022424',
                     'E0022425',
                     'E0022426',
                     'E0022427',
                     'E0022428',
                     'E0022429',
                     'E0022430',
                     'E0022431',
                     'E0022432',
                     'E0022433',
                     'E0022434',
                     'E0022435',
                     'E0022436',
                     'E0022437',
                     'E0022438',
                     'E0022439',
                     'E0022440',
                     'E0022441',
                     'E0022442',
                     'E0022443',
                     'E0022444',
                     'E0022445',
                     'E0022446',
                     'E0022447',
                     'E0022448',
                     'E0022449',
                     'E0022450',
                     'E0022451',
                     'E0022452',
                     'E0022501',
                     'E0022502',
                     'E0022503',
                     'E0022504',
                     'E0022505',
                     'E0022506',
                     'E0022507',
                     'E0022508',
                     'E0022509',
                     'E0022510',
                     'E0022511',
                     'E0022512',
                     'E0022513',
                     'E0022514',
                     'E0022515',
                     'E0022516',
                     'E0022517',
                     'E0022518',
                     'E0022519',
                     'E0022520',
                     'E0022521',
                     'E0022522',
                     'E0022523',
                     'E0022524',
                     'E0022525',
                     'E0022526',
                     'E0022527',
                     'E0022528',
                     'E0022529',
                     'E0022530',
                     'E0022531',
                     'E0022532',
                     'E0022533',
                     'E0022534',
                     'E0022535',
                     'E0022536',
                     'E0022537',
                     'E0022538',
                     'E0022539',
                     'E0022540',
                     'E0022541',
                     'E0022542',
                     'E0022543',
                     'E0022544',
                     'E0022545',
                     'E0022546',
                     'E0022547',
                     'E0022548',
                     'E0022549',
                     'E0022550',
                     'E0022551',
                     'E0022552',
                     'E0022553',
                     'E0022601',
                     'E0022602',
                     'E0022603',
                     'E0022604',
                     'E0022605',
                     'E0022606',
                     'E0022607',
                     'E0022608',
                     'E0022609',
                     'E0022610',
                     'E0022611',
                     'E0022612',
                     'E0022613',
                     'E0022614',
                     'E0022615',
                     'E0022616',
                     'E0022617',
                     'E0022618',
                     'E0022619',
                     'E0022620',
                     'E0022621',
                     'E0022622',
                     'E0022623',
                     'E0022624',
                     'E0022625',
                     'E0022626',
                     'E0022627',
                     'E0022628',
                     'E0022629',
                     'E0022630',
                     'E0022631',
                     'E0022632',
                     'E0022633',
                     'E0022634',
                     'E0022635',
                     'E0022636',
                     'E0022637',
                     'E0022638',
                     'E0022639',
                     'E0022640',
                     'E0022641',
                     'E0022642',
                     'E0022643',
                     'E0022644',
                     'E0022645',
                     'E0022646',
                     'E0022647',
                     'E0022648',
                     'E0022649',
                     'E0022650',
                     'E0022651',
                     'E0022652',
                     'E0022701',
                     'E0022702',
                     'E0022703',
                     'E0022704',
                     'E0022705',
                     'E0022706',
                     'E0022707',
                     'E0022708',
                     'E0022709',
                     'E0022710',
                     'E0022711',
                     'E0022712',
                     'E0022713',
                     'E0022714',
                     'E0022715',
                     'E0022716',
                     'E0022717',
                     'E0022718',
                     'E0022719',
                     'E0022720',
                     'E0022721',
                     'E0022722',
                     'E0022723',
                     'E0022724',
                     'E0022725',
                     'E0022726',
                     'E0022727',
                     'E0022728',
                     'E0022729',
                     'E0022730',
                     'E0022731',
                     'E0022732',
                     'E0022733',
                     'E0022734',
                     'E0022735',
                     'E0022736',
                     'E0022737',
                     'E0022738',
                     'E0022739',
                     'E0022740',
                     'E0022741',
                     'E0022742',
                     'E0022743',
                     'E0022744',
                     'E0022745',
                     'E0022746',
                     'E0022747',
                     'E0022748',
                     'E0022749',
                     'E0022750',
                     'E0022751',
                     'E0022752',
                     'E0022801',
                     'E0022802',
                     'E0022803',
                     'E0022804',
                     'E0022805',
                     'E0022806',
                     'E0022807',
                     'E0022808',
                     'E0022809',
                     'E0022810',
                     'E0022811',
                     'E0022812',
                     'E0022813',
                     'E0022814',
                     'E0022815',
                     'E0022816',
                     'E0022817',
                     'E0022818',
                     'E0022819',
                     'E0022820',
                     'E0022821',
                     'E0022822',
                     'E0022823',
                     'E0022824',
                     'E0022825',
                     'E0022826',
                     'E0022827',
                     'E0022828',
                     'E0022829',
                     'E0022830',
                     'R0000100',
                     'R0489900',
                     'R0490000',
                     'R0490100',
                     'R0490200',
                     'R0490300',
                     'R0490900',
                     'R0491100',
                     'R0491200',
                     'R0536300',
                     'R0536401',
                     'R0536402',
                     'R0609500',
                     'R0609600',
                     'R0609700',
                     'R0609800',
                     'R0609900',
                     'R0610400',
                     'R0610500',
                     'R0610600',
                     'R0688100',
                     'R0688200',
                     'R0688300',
                     'R1201500',
                     'R1201600',
                     'R1204500',
                     'R1204600',
                     'R1209101',
                     'R1209102',
                     'R1209103',
                     'R1209104',
                     'R1209105',
                     'R1209106',
                     'R1209107',
                     'R1210200',
                     'R1211100',
                     'R1235800',
                     'R1236100',
                     'R1236101',
                     'R1236201',
                     'R1482600',
                     'R2340900',
                     'R2341000',
                     'R2341100',
                     'R2341200',
                     'R2341300',
                     'R2341900',
                     'R2342100',
                     'R2342200',
                     'R2399400',
                     'R2469000',
                     'R2469100',
                     'R2469200',
                     'R2469300',
                     'R2469400',
                     'R2469500',
                     'R2469600',
                     'R2469700',
                     'R2469800',
                     'R2563300',
                     'R2568001',
                     'R2568002',
                     'R2568003',
                     'R2568004',
                     'R2568005',
                     'R2568006',
                     'R2568007',
                     'R2568008',
                     'R2568009',
                     'R2569300',
                     'R2570100',
                     'R2600300',
                     'R2600301',
                     'R2600401',
                     'R3649900',
                     'R3650000',
                     'R3650100',
                     'R3650200',
                     'R3650300',
                     'R3650900',
                     'R3651100',
                     'R3651200',
                     'R3664100',
                     'R3664200',
                     'R3664300',
                     'R3707300',
                     'R3782800',
                     'R3782900',
                     'R3783000',
                     'R3783100',
                     'R3783200',
                     'R3783300',
                     'R3783400',
                     'R3783500',
                     'R3783600',
                     'R3884900',
                     'R3889701',
                     'R3889702',
                     'R3889703',
                     'R3889704',
                     'R3889705',
                     'R3889706',
                     'R3889707',
                     'R3889708',
                     'R3889709',
                     'R3891300',
                     'R3892100',
                     'R3923700',
                     'R3923701',
                     'R3958500',
                     'R3958501',
                     'R5098600',
                     'R5098700',
                     'R5098800',
                     'R5098900',
                     'R5099000',
                     'R5099600',
                     'R5099800',
                     'R5099900',
                     'R5113200',
                     'R5113300',
                     'R5113400',
                     'R5164800',
                     'R5305200',
                     'R5305300',
                     'R5305400',
                     'R5305500',
                     'R5305600',
                     'R5305700',
                     'R5305800',
                     'R5305900',
                     'R5306000',
                     'R5464100',
                     'R5471801',
                     'R5471802',
                     'R5471803',
                     'R5471804',
                     'R5471805',
                     'R5471806',
                     'R5471807',
                     'R5471808',
                     'R5471809',
                     'R5473300',
                     'R5474100',
                     'R5510600',
                     'R5510700',
                     'R5528900',
                     'R5529000',
                     'R6827200',
                     'R6827300',
                     'R6827400',
                     'R6827500',
                     'R6827600',
                     'R6828200',
                     'R6828400',
                     'R6828500',
                     'R6841900',
                     'R6842000',
                     'R6842100',
                     'R6893100',
                     'R7047000',
                     'R7047100',
                     'R7047200',
                     'R7047300',
                     'R7047400',
                     'R7047500',
                     'R7047600',
                     'R7047700',
                     'R7227800',
                     'R7235601',
                     'R7235602',
                     'R7235603',
                     'R7235604',
                     'R7235605',
                     'R7235606',
                     'R7235607',
                     'R7235608',
                     'R7237000',
                     'R7237800',
                     'R7274200',
                     'R7274300',
                     'S1055500',
                     'S1055600',
                     'S1055700',
                     'S1055800',
                     'S1055900',
                     'S1056800',
                     'S1057000',
                     'S1057100',
                     'S1070400',
                     'S1070500',
                     'S1070600',
                     'S1516100',
                     'S1516200',
                     'S1516300',
                     'S1516400',
                     'S1516500',
                     'S1516600',
                     'S1516700',
                     'S1516800',
                     'S1516900',
                     'S1517000',
                     'S1517100',
                     'S1541700',
                     'S1550401',
                     'S1550402',
                     'S1550403',
                     'S1550404',
                     'S1550405',
                     'S1550406',
                     'S1550407',
                     'S1550408',
                     'S1550409',
                     'S1550410',
                     'S1550411',
                     'S1552200',
                     'S1553100',
                     'S1598100',
                     'S1598200',
                     'S1601900',
                     'S1602000',
                     'S1602100',
                     'S1602200',
                     'S1602300',
                     'S1602400',
                     'S1602500',
                     'S1602600',
                     'S1602700',
                     'S1602800',
                     'S1602900',
                     'S1603000',
                     'S1603100',
                     'S1603200',
                     'S1603300',
                     'S1603400',
                     'S1603500',
                     'S1603600',
                     'S1603700',
                     'S1603800',
                     'S1603900',
                     'S1604000',
                     'S2011500',
                     'S2020301',
                     'S2020302',
                     'S2020303',
                     'S2020304',
                     'S2020305',
                     'S2020306',
                     'S2020307',
                     'S2020308',
                     'S2020309',
                     'S2020310',
                     'S2022000',
                     'S2022700',
                     'S2067000',
                     'S2067100',
                     'S3134400',
                     'S3134500',
                     'S3134600',
                     'S3134700',
                     'S3135500',
                     'S3135700',
                     'S3135800',
                     'S3148600',
                     'S3148700',
                     'S3148800',
                     'S3582600',
                     'S3582700',
                     'S3582800',
                     'S3582900',
                     'S3583000',
                     'S3583100',
                     'S3583200',
                     'S3583300',
                     'S3583400',
                     'S3583500',
                     'S3658000',
                     'S3658100',
                     'S3658200',
                     'S3658300',
                     'S3658400',
                     'S3658500',
                     'S3658600',
                     'S3659000',
                     'S3659100',
                     'S3659200',
                     'S3659300',
                     'S3659400',
                     'S3659500',
                     'S3659600',
                     'S3679000',
                     'S3679100',
                     'S3679200',
                     'S3679300',
                     'S3679400',
                     'S3679500',
                     'S3679600',
                     'S3679700',
                     'S3679800',
                     'S3681000',
                     'S3681100',
                     'S3681200',
                     'S3681300',
                     'S3681400',
                     'S3681500',
                     'S3681600',
                     'S3681700',
                     'S3681800',
                     'S3695000',
                     'S3695100',
                     'S3695200',
                     'S3695300',
                     'S3695400',
                     'S3695500',
                     'S3695600',
                     'S3695700',
                     'S3695800',
                     'S3697000',
                     'S3697100',
                     'S3697200',
                     'S3697300',
                     'S3697400',
                     'S3697500',
                     'S3697600',
                     'S3697700',
                     'S3697800',
                     'S3711000',
                     'S3711100',
                     'S3711200',
                     'S3711300',
                     'S3711400',
                     'S3711500',
                     'S3711600',
                     'S3711700',
                     'S3711800',
                     'S3713000',
                     'S3713100',
                     'S3713200',
                     'S3713300',
                     'S3713400',
                     'S3713500',
                     'S3713600',
                     'S3713700',
                     'S3713800',
                     'S3727000',
                     'S3727100',
                     'S3727200',
                     'S3727300',
                     'S3727400',
                     'S3727500',
                     'S3727600',
                     'S3727700',
                     'S3729000',
                     'S3729100',
                     'S3729200',
                     'S3729300',
                     'S3729400',
                     'S3729500',
                     'S3729600',
                     'S3729700',
                     'S3755000',
                     'S3755100',
                     'S3755200',
                     'S3755300',
                     'S3755400',
                     'S3755500',
                     'S3755600',
                     'S3755700',
                     'S3755800',
                     'S3755900',
                     'S3757000',
                     'S3757100',
                     'S3757200',
                     'S3757300',
                     'S3757400',
                     'S3757500',
                     'S3757600',
                     'S3757700',
                     'S3757800',
                     'S3757900',
                     'S3812400',
                     'S3821501',
                     'S3821502',
                     'S3821503',
                     'S3821504',
                     'S3821505',
                     'S3821506',
                     'S3821507',
                     'S3822900',
                     'S3823600',
                     'S3861600',
                     'S3861700',
                     'S4799400',
                     'S4799500',
                     'S4799600',
                     'S4799700',
                     'S4800500',
                     'S4800700',
                     'S4800800',
                     'S4806500',
                     'S4806600',
                     'S4806700',
                     'S5041000',
                     'S5041100',
                     'S5041200',
                     'S5041300',
                     'S5041400',
                     'S5041500',
                     'S5041600',
                     'S5041700',
                     'S5041800',
                     'S5041900',
                     'S5042000',
                     'S5042100',
                     'S5042200',
                     'S5042300',
                     'S5236800',
                     'S5236900',
                     'S5237000',
                     'S5237100',
                     'S5237200',
                     'S5237300',
                     'S5237400',
                     'S5412800',
                     'S5421801',
                     'S5421802',
                     'S5421803',
                     'S5421804',
                     'S5421805',
                     'S5421806',
                     'S5421807',
                     'S5421808',
                     'S5421809',
                     'S5423000',
                     'S5423600',
                     'S5444200',
                     'S5444300',
                     'S6500800',
                     'S6500900',
                     'S6501000',
                     'S6501100',
                     'S6501800',
                     'S6502200',
                     'S6502300',
                     'S6508000',
                     'S6508100',
                     'S6508200',
                     'S6782200',
                     'S6782300',
                     'S6782400',
                     'S6782500',
                     'S6782600',
                     'S6782700',
                     'S6782800',
                     'S6782900',
                     'S6783000',
                     'S6783100',
                     'S6783200',
                     'S6783300',
                     'S6783400',
                     'S6783500',
                     'S6783600',
                     'S6783700',
                     'S6783800',
                     'S6783900',
                     'S7026600',
                     'S7026700',
                     'S7026800',
                     'S7026900',
                     'S7027000',
                     'S7027100',
                     'S7027200',
                     'S7027300',
                     'S7027400',
                     'S7513700',
                     'S7523100',
                     'S7523200',
                     'S7523300',
                     'S7523400',
                     'S7523500',
                     'S7523600',
                     'S7523700',
                     'S7523800',
                     'S7523900',
                     'S7525100',
                     'S7525700',
                     'S7545300',
                     'S7545400',
                     'S7545500',
                     'S7545600',
                     'S8496300',
                     'S8496400',
                     'S8496500',
                     'S8496600',
                     'S8497400',
                     'S8497900',
                     'S8498000',
                     'S8504000',
                     'S8504100',
                     'S8504200',
                     'S8688800',
                     'S8688900',
                     'S8689000',
                     'S8689100',
                     'S8689200',
                     'S8689300',
                     'S8689400',
                     'S8689500',
                     'S8689600',
                     'S8689700',
                     'S8689800',
                     'S8689900',
                     'S8690000',
                     'S8690100',
                     'S8690200',
                     'S8690300',
                     'S8690400',
                     'S8690500',
                     'S8997800',
                     'S8997900',
                     'S8998000',
                     'S8998100',
                     'S8998200',
                     'S8998300',
                     'S8998400',
                     'S8998500',
                     'S8998600',
                     'T0014100',
                     'T0023600',
                     'T0023700',
                     'T0023800',
                     'T0023900',
                     'T0024000',
                     'T0024100',
                     'T0024200',
                     'T0024300',
                     'T0025400',
                     'T0026000',
                     'T0042100',
                     'T0042200',
                     'T0889600',
                     'T0889700',
                     'T0889800',
                     'T0889900',
                     'T0890700',
                     'T0891100',
                     'T0891200',
                     'T0897100',
                     'T0897200',
                     'T0897300',
                     'T1108600',
                     'T1108700',
                     'T1108800',
                     'T1108900',
                     'T1109000',
                     'T1109100',
                     'T1109200',
                     'T1109300',
                     'T1109400',
                     'T1109500',
                     'T1109600',
                     'T1109700',
                     'T1109800',
                     'T1109900',
                     'T1110000',
                     'T1110100',
                     'T1407700',
                     'T1407800',
                     'T1407900',
                     'T1408000',
                     'T1408100',
                     'T1408200',
                     'T1408300',
                     'T1408400',
                     'T2016200',
                     'T2018500',
                     'T2018600',
                     'T2018700',
                     'T2018800',
                     'T2018900',
                     'T2019000',
                     'T2019100',
                     'T2019200',
                     'T2020300',
                     'T2020900',
                     'T2022500',
                     'T2022600',
                     'T3002800',
                     'T3002900',
                     'T3003000',
                     'T3003100',
                     'T3003900',
                     'T3004300',
                     'T3004400',
                     'T3010400',
                     'T3010500',
                     'T3010600',
                     'T3186000',
                     'T3186100',
                     'T3186200',
                     'T3186300',
                     'T3186400',
                     'T3186500',
                     'T3186600',
                     'T3186700',
                     'T3186800',
                     'T3186900',
                     'T3187000',
                     'T3187100',
                     'T3187200',
                     'T3187300',
                     'T3187400',
                     'T3187500',
                     'T3187600',
                     'T3187700',
                     'T3504400',
                     'T3504500',
                     'T3504600',
                     'T3504700',
                     'T3504800',
                     'T3504900',
                     'T3505000',
                     'T3505100',
                     'T3606500',
                     'T3609000',
                     'T3609100',
                     'T3609200',
                     'T3609300',
                     'T3609400',
                     'T3609500',
                     'T3609600',
                     'T3609700',
                     'T3609800',
                     'T3611000',
                     'T3611600',
                     'T3613300',
                     'T3613400',
                     'T4405800',
                     'T4405900',
                     'T4406000',
                     'T4406100',
                     'T4406900',
                     'T4407300',
                     'T4407400',
                     'T4411200',
                     'T4411300',
                     'T4411400',
                     'T4596900',
                     'T4597000',
                     'T4597100',
                     'T4597200',
                     'T4597300',
                     'T4597400',
                     'T4597500',
                     'T4597600',
                     'T4597700',
                     'T4597800',
                     'T4597900',
                     'T4598000',
                     'T4598100',
                     'T4598200',
                     'T4598300',
                     'T4598400',
                     'T4598500',
                     'T4598600',
                     'T4934100',
                     'T4934200',
                     'T4934300',
                     'T4934400',
                     'T4934500',
                     'T4934600',
                     'T4934700',
                     'T4934800',
                     'T4934900',
                     'T5206900',
                     'T5209400',
                     'T5209500',
                     'T5209600',
                     'T5209700',
                     'T5209800',
                     'T5209900',
                     'T5210000',
                     'T5210100',
                     'T5210200',
                     'T5211400',
                     'T5211700',
                     'T5213200',
                     'T5213300',
                     'T6055300',
                     'T6055400',
                     'T6055500',
                     'T6055600',
                     'T6056400',
                     'T6056800',
                     'T6056900',
                     'T6061300',
                     'T6061400',
                     'T6061500',
                     'T6230100',
                     'T6230200',
                     'T6230300',
                     'T6230400',
                     'T6230500',
                     'T6230600',
                     'T6230700',
                     'T6230800',
                     'T6230900',
                     'T6231000',
                     'T6231100',
                     'T6231200',
                     'T6231300',
                     'T6231400',
                     'T6231500',
                     'T6231600',
                     'T6231700',
                     'T6231800',
                     'T6587300',
                     'T6587400',
                     'T6587500',
                     'T6587600',
                     'T6587700',
                     'T6587800',
                     'T6587900',
                     'T6588000',
                     'T6588100',
                     'T6656700',
                     'T6660000',
                     'T6660100',
                     'T6660200',
                     'T6660300',
                     'T6660400',
                     'T6660500',
                     'T6660600',
                     'T6660700',
                     'T6660800',
                     'T6660900',
                     'T6661000',
                     'T6661100',
                     'T6661200',
                     'T6662800',
                     'T6663100',
                     'T6665000',
                     'T6665100',
                     'T7545400',
                     'T7545500',
                     'T7545600',
                     'T7545700',
                     'T7546500',
                     'T7546900',
                     'T7547000',
                     'T7551400',
                     'T7551500',
                     'T7551600',
                     'T7731100',
                     'T7731200',
                     'T7731300',
                     'T7731400',
                     'T7731500',
                     'T7731600',
                     'T7731700',
                     'T7731800',
                     'T7731900',
                     'T7732000',
                     'T7732100',
                     'T7732200',
                     'T7732300',
                     'T7732400',
                     'T7732500',
                     'T7732600',
                     'T7732700',
                     'T7732800',
                     'T7732900',
                     'T7733000',
                     'T8093500',
                     'T8093600',
                     'T8093700',
                     'T8093800',
                     'T8093900',
                     'T8094000',
                     'T8094100',
                     'T8094200',
                     'T8094300',
                     'T8094400',
                     'T8094500',
                     'T8094600',
                     'T8094700',
                     'T8129100',
                     'T8131800',
                     'T8131900',
                     'T8132000',
                     'T8132100',
                     'T8132200',
                     'T8132300',
                     'T8132400',
                     'T8132500',
                     'T8132600',
                     'T8132700',
                     'T8134000',
                     'T8134300',
                     'T8135900',
                     'T8136000',
                     'T8976500',
                     'T8976600',
                     'T8976700',
                     'T8976800',
                     'T8977600',
                     'T8978000',
                     'T8978100',
                     'T9132600',
                     'T9132700',
                     'T9132800',
                     'T9132900',
                     'T9133000',
                     'T9133100',
                     'T9133200',
                     'T9133300',
                     'T9133400',
                     'T9133500',
                     'T9133600',
                     'T9133700',
                     'T9133800',
                     'T9133900',
                     'T9134000',
                     'T9134100',
                     'T9134200',
                     'T9134300',
                     'T9336000',
                     'T9336100',
                     'T9336200',
                     'T9336300',
                     'T9336400',
                     'T9336500',
                     'T9336600',
                     'T9336700',
                     'T9336800',
                     'T9336900',
                     'U0008900',
                     'U0011900',
                     'U0012000',
                     'U0012100',
                     'U0012200',
                     'U0012300',
                     'U0012400',
                     'U0012500',
                     'U0012600',
                     'U0012700',
                     'U0012800',
                     'U0012900',
                     'U0013000',
                     'U0014500',
                     'U0014800',
                     'U0017100',
                     'U0017200',
                     'U0956700',
                     'U0956800',
                     'U0956900',
                     'U0957000',
                     'U0957800',
                     'U0958200',
                     'U0958300',
                     'U1125900',
                     'U1126000',
                     'U1126100',
                     'U1126200',
                     'U1126300',
                     'U1126400',
                     'U1126500',
                     'U1126600',
                     'U1126700',
                     'U1126800',
                     'U1126900',
                     'U1127000',
                     'U1127100',
                     'U1127200',
                     'U1127300',
                     'U1127400',
                     'U1127500',
                     'U1127600',
                     'U1127700',
                     'U1127800',
                     'U1127900',
                     'U1128000',
                     'U1128100',
                     'U1128200',
                     'U1350100',
                     'U1350200',
                     'U1350300',
                     'U1350400',
                     'U1350500',
                     'U1350600',
                     'U1350700',
                     'U1350800',
                     'U1350900',
                     'U1351000',
                     'U1351100',
                     'U1351200',
                     'Z9083900',
                     'Z9084100',
                     'Z9084200',
                     'Z9084300',
                     'Z9084400',
                     'Z9084500',
                     'Z9084600',
                     'Z9084700')


# Handle missing values

data_97[data_97 == -1] = NA  # Refused 
data_97[data_97 == -2] = NA  # Dont know 
data_97[data_97 == -3] = NA  # Invalid missing 
data_97[data_97 == -4] = NA  # Valid missing 
data_97[data_97 == -5] = NA  # Non-interview
